apps/video-intelligence/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
pymongo==4.6.0
voyageai==0.2.3
openai==1.3.8
python-dotenv==1.0.0
websockets==12.0
numpy==1.24.3
pathlib==1.0.1
aiofiles==23.2.1
